Bugfix: a hot fix of atom.connect()

+++ enzy_htp/structure/structure.py
@@ -252,7 +252,7 @@ class Structure(DoubleLinkedNode):
                 result.extend(residue.atoms)
         return result
 
-    def find_atoms_in_range(self, center: Union[Atom, Tuple[int, int, int]],
+    def find_atoms_in_range(self, center: Union[Atom, Tuple[float, float, float]],
                             range_distance: float) -> List[Atom]:
         """find atoms in {range} of {center}. return a list of atoms found"""
         result = []
